windows: fix Visual Studio warns that not all control paths return
Which is not true.  On the other hand Clang warns that I am stupid for
being over-exaustive on this switch case if I add a "default" branch.
Do cross-platform development with taliban warning levels enabled on
all your compilers at the same time and feel the unique experience of
being Poland in the WWII.