Don't include the Rhea buildsystem in ours.
The rhea buildsystem is not designed to be included like that.

It populates the standard CMAKE_CXX_FLAGS variable for example
which might not be compatible with ours.  It also adds some
option() entries to the cache such as 'BUILD_DOCUMENTATION', generic
names which appear in the buildsystem of cassowary, even though
they relate to rhea.

Attempt to find rhea using standard CMake finder commands.  If that
fails, use the CMake ExternalProject module to build and install a
local copy of rhea.  Create an IMPORTED target for Rhea for user
convenience. Ideally something similar would be provided by rhea,
but that is not currently the case.