Commit before rendering so that there's less lag
The previous approch of queueing for the next iteration of the main loop
would cause the solver to sometimes lag one frame behind.

Now we use override the `updatePolish()` function such that the
constraints are update just before the system is about to render.  This
ensures that the updates happen before rendering and that no more than
once per frame we update -- unless, of course, commit is called
manually.