Fix failure to exchange mutually exclusive constraints
Exchanging mutually exclusive constraints would fail non
deterministically.  The test added to `tst_Constraint.qml` explains
this situation.

The approach to solve this issue is to divide the application of
changes into several phases. All *removals* are done first and later
all *additions*.  Objects have to be tracked because now the sequence
*add(C)*, *remove(C)* would be reversed having unexpected results --
instead we completelly cancel the sequence in this case, under the
assumption that the effect of that sequence is indeed a no-op.

It is hard to properly track constraints, because Rhea constraints do
not have value semantics and it is too hard to build these semantics
on top. The lack of value semantics created problems with the way we
were changing weights and strengths (by re-adding the contraint with
different strenght/weight).  We now use the `change_{strength|weight}`
methods in Rhea.  These do not allow to change the strength of a
required constraint.  We are ok with this limitation for now until we
find a impeeded core use-case.

Finally, the code has become a bit clumsy after this refactor.  A
further refactor that I would like to do in the future is to stop
manipulating the `simplex_solver` directly on all the individual objects
of the `Contextual`.  Instead, they should just call explicit methods in
the context that should internally encapsulate all the deferring and
logistics of managing a consistent state in the solver.