refactor: move all calls into rhea to the Context object
This means that we do not distribute the responsiblity of updating the
solver among all objects anymore.  Instead, when an object wants to
change the solver, it makes a call in the `Context` specifying the
specific operation that it wants to do.  In this way, the context has
more knowledge about how to order things, how to optimize away
constraint changes and when to call `solve()` and `resolve()`.

A couple of bugs have been fixed with this commit, as illustrated by the
added unit tests.

Plus, since this commit the whole program does not crash anymore when
rhea throws an error.  Instead, the error is just logged.