Prevent infinite loops when contraints depend on variables
If constraints when property or edit suggestions depend on the values of
variables, it is possible for the system to go into infinite loops.
This is so because slight rounding errors might cause the variable to
infinitelly change even though the program is logically correct.  I
sadly could not reproduce this situation in a test but have definitelly
got into that in example cases.  `rhea::near_zero` is safe since it
actually literally defines the rhea rounding error.