example: Add example of draggable squares
The squares may block or push eachother depending on whether the option
is enabled.

This example pushes the system to its limits because it has contraints
that are enabled or disabled depending on the values that are mentioned
in the constraints themselves.

Also, it seems to show that stays are weird and sometimes it is better
to just feed back the values through an edit.