Replace stays by edits that feedback the target's value
Stays seem to be broken in unpredictable ways.  This problem sadly
happens more often than we would like with Aqt.Cassowary, yet it is hard
to reproduce in a test since it is rather non-deterministic.  The
problem can be followed here:

    https://github.com/Nocte-/rhea/issues/40

However, the change in this commit allows to solve the problem in
practice for Aqt.Cassowary.  All tests still pass with the new way to
define stays, so it seems a workable solution until the underlying
problems are solved in Rhea.